requirements.txt
python-telegram-bot==20.7
geopy==2.4.1
tabulate==0.9.0 